chore: Update catalog service structure
Change file structure of the catalog service backend to be in line with
other crates
